Follow the review recommendations: increase presicion from float to double, using vmData instead of vmSize, if command is greater that 50 characters truncate the name and display ...

+++ include/processor.h
@@ -13,7 +13,7 @@ using LinuxParser::CPUStates;
 
 class Processor {
  public:
-  float Utilization();  // TODO: See src/processor.cpp
+  double Utilization();  // TODO: See src/processor.cpp
 
   // TODO: Declare any necessary private members
  private:
